requirements.txt
python-kasa
pillow
mss
# Optional (for improved screen capture averaging)
dxcam
numpy
